+++ lib/code_format.dart
@@ -1,6 +1,6 @@
+import 'package:flutter/foundation.dart';
 
-enum CodeFormat
-{
+enum CodeFormat {
   /// Aztec 2D barcode format.
   AZTEC,
 
@@ -25,9 +25,6 @@ enum CodeFormat
   /// EAN-13 1D format.
   EAN_13,
 
-  /// ITF (Interleaved Two of Five) 1D format.
-  ITF,
-
   /// MaxiCode 2D barcode format.
   MAXICODE,
 
@@ -48,112 +45,232 @@ enum CodeFormat
 
   /// UPC-E 1D format.
   UPC_E,
-
-  /// UPC/EAN extension format. Not a stand-alone format.
-  UPC_EAN_EXTENSION,
 }
 
-class CodeFormatUtils
-{
+extension CodeFormatUtils on CodeFormat {
   static final String _CODE_FORMAT_PROPERTY_AZTEC_ENABLED = "DEC_AZTEC_ENABLED";
-  static final String _CODE_FORMAT_PROPERTY_CODABAR_ENABLED = "DEC_CODABAR_ENABLED";
-  static final String _CODE_FORMAT_PROPERTY_CODE_39_ENABLED = "DEC_CODE39_ENABLED";
-  static final String _CODE_FORMAT_PROPERTY_CODE_93_ENABLED = "DEC_CODE93_ENABLED";
-  static final String _CODE_FORMAT_PROPERTY_CODE_128_ENABLED = "DEC_CODE128_ENABLED";
-  static final String _CODE_FORMAT_PROPERTY_DATAMATRIX_ENABLED = "DEC_DATAMATRIX_ENABLED";
+  static final String _CODE_FORMAT_PROPERTY_CODABAR_ENABLED =
+      "DEC_CODABAR_ENABLED";
+  static final String _CODE_FORMAT_PROPERTY_CODE_39_ENABLED =
+      "DEC_CODE39_ENABLED";
+  static final String _CODE_FORMAT_PROPERTY_CODE_93_ENABLED =
+      "DEC_CODE93_ENABLED";
+  static final String _CODE_FORMAT_PROPERTY_CODE_128_ENABLED =
+      "DEC_CODE128_ENABLED";
+  static final String _CODE_FORMAT_PROPERTY_DATAMATRIX_ENABLED =
+      "DEC_DATAMATRIX_ENABLED";
   static final String _CODE_FORMAT_PROPERTY_EAN_8_ENABLED = "DEC_EAN8_ENABLED";
-  static final String _CODE_FORMAT_PROPERTY_EAN_13_ENABLED = "DEC_EAN13_ENABLED";
-  static final String _CODE_FORMAT_PROPERTY_MAXICODE_ENABLED = "DEC_MAXICODE_ENABLED";
-  static final String _CODE_FORMAT_PROPERTY_PDF_417_ENABLED = "DEC_PDF417_ENABLED";
+  static final String _CODE_FORMAT_PROPERTY_EAN_13_ENABLED =
+      "DEC_EAN13_ENABLED";
+  static final String _CODE_FORMAT_PROPERTY_MAXICODE_ENABLED =
+      "DEC_MAXICODE_ENABLED";
+  static final String _CODE_FORMAT_PROPERTY_PDF_417_ENABLED =
+      "DEC_PDF417_ENABLED";
   static final String _CODE_FORMAT_PROPERTY_QR_CODE_ENABLED = "DEC_QR_ENABLED";
   static final String _CODE_FORMAT_PROPERTY_RSS_ENABLED = "DEC_RSS_14_ENABLED";
-  static final String _CODE_FORMAT_PROPERTY_RSS_EXPANDED_ENABLED = "DEC_RSS_EXPANDED_ENABLED";
+  static final String _CODE_FORMAT_PROPERTY_RSS_EXPANDED_ENABLED =
+      "DEC_RSS_EXPANDED_ENABLED";
   static final String _CODE_FORMAT_PROPERTY_UPC_A_ENABLE = "DEC_UPCA_ENABLE";
   static final String _CODE_FORMAT_PROPERTY_UPC_E_ENABLED = "DEC_UPCE0_ENABLED";
 
-  CodeFormatUtils.get();
+  /// All supported code formats
+  static const List<CodeFormat> ALL_FORMATS = CodeFormat.values;
 
-  static final _mapValueOfName = {
-  "AZTEC": CodeFormat.AZTEC,
-  "CODABAR": CodeFormat.CODABAR,
-  "CODE_39": CodeFormat.CODE_39,
-  "CODE_93": CodeFormat.CODE_93,
-  "CODE_128": CodeFormat.CODE_128,
-  "DATA_MATRIX": CodeFormat.DATA_MATRIX,
-  "EAN_8": CodeFormat.EAN_8,
-  "EAN_13": CodeFormat.EAN_13,
-  "ITF": CodeFormat.ITF,
-  "MAXICODE": CodeFormat.MAXICODE,
-  "PDF_417": CodeFormat.PDF_417,
-  "QR_CODE": CodeFormat.QR_CODE,
-  "RSS_14": CodeFormat.RSS_14,
-  "RSS_EXPANDED": CodeFormat.RSS_EXPANDED,
-  "UPC_A": CodeFormat.UPC_A,
-  "UPC_E": CodeFormat.UPC_E,
-  "UPC_EAN_EXTENSION": CodeFormat.UPC_EAN_EXTENSION,
-  };
+  /// One dimensional code formats
+  static const List<CodeFormat> ALL_1D_FORMATS = [
+    CodeFormat.CODABAR,
+    CodeFormat.CODE_39,
+    CodeFormat.CODE_93,
+    CodeFormat.CODE_128,
+    CodeFormat.EAN_8,
+    CodeFormat.EAN_13,
+    CodeFormat.RSS_14,
+    CodeFormat.RSS_EXPANDED,
+    CodeFormat.UPC_A,
+    CodeFormat.UPC_E,
+  ];
 
-  CodeFormat valueOf(String name) {
-    CodeFormat value = null;
-    try
-    {
-      value = _mapValueOfName[name];
-    }
-    catch(e)
-    {
-      print(e);
-    }
-    return value;
-  }
+  /// One dimensional product code formats
+  static const List<CodeFormat> ALL_1D_PRODUCT_FORMATS = [
+    CodeFormat.EAN_8,
+    CodeFormat.EAN_13,
+    CodeFormat.UPC_A,
+    CodeFormat.UPC_E,
+  ];
 
-  String getPropertyName(CodeFormat value) {
-    if(value == null) return null;
-    try
-    {
-      switch(value){
-        case CodeFormat.AZTEC: return _CODE_FORMAT_PROPERTY_AZTEC_ENABLED;
-        case CodeFormat.CODABAR: return _CODE_FORMAT_PROPERTY_CODABAR_ENABLED;
-        case CodeFormat.CODE_39:return _CODE_FORMAT_PROPERTY_CODE_39_ENABLED;
-        case CodeFormat.CODE_93:return _CODE_FORMAT_PROPERTY_CODE_93_ENABLED;
-        case CodeFormat.CODE_128:return _CODE_FORMAT_PROPERTY_CODE_128_ENABLED;
-        case CodeFormat.DATA_MATRIX:return _CODE_FORMAT_PROPERTY_DATAMATRIX_ENABLED;
-        case CodeFormat.EAN_8:return _CODE_FORMAT_PROPERTY_EAN_8_ENABLED;
-        case CodeFormat.EAN_13:return _CODE_FORMAT_PROPERTY_EAN_13_ENABLED;
-        case CodeFormat.ITF:break;
-        case CodeFormat.MAXICODE:return _CODE_FORMAT_PROPERTY_MAXICODE_ENABLED;
-        case CodeFormat.PDF_417:return _CODE_FORMAT_PROPERTY_PDF_417_ENABLED;
-        case CodeFormat.QR_CODE:return _CODE_FORMAT_PROPERTY_QR_CODE_ENABLED;
-        case CodeFormat.RSS_14:return _CODE_FORMAT_PROPERTY_RSS_ENABLED;
-        case CodeFormat.RSS_EXPANDED:return _CODE_FORMAT_PROPERTY_RSS_EXPANDED_ENABLED;
-        case CodeFormat.UPC_A:return _CODE_FORMAT_PROPERTY_UPC_A_ENABLE;
-        case CodeFormat.UPC_E:return _CODE_FORMAT_PROPERTY_UPC_E_ENABLED;
-        case CodeFormat.UPC_EAN_EXTENSION:break;
+  /// One dimensional industrial code formats
+  static const List<CodeFormat> ALL_1D_INDUSTRIAL_FORMATS = [
+    CodeFormat.CODABAR,
+    CodeFormat.CODE_39,
+    CodeFormat.CODE_93,
+    CodeFormat.CODE_128,
+    CodeFormat.RSS_14,
+    CodeFormat.RSS_EXPANDED
+  ];
+
+  /// Two dimensional code formats
+  static const List<CodeFormat> ALL_2D_FORMATS = [
+    CodeFormat.AZTEC,
+    CodeFormat.DATA_MATRIX,
+    CodeFormat.MAXICODE,
+    CodeFormat.PDF_417,
+    CodeFormat.QR_CODE
+  ];
+
+  String get name => describeEnum(this);
+
+  String get propertyName {
+    try {
+      switch (this) {
+        case CodeFormat.AZTEC:
+          return _CODE_FORMAT_PROPERTY_AZTEC_ENABLED;
+        case CodeFormat.CODABAR:
+          return _CODE_FORMAT_PROPERTY_CODABAR_ENABLED;
+        case CodeFormat.CODE_39:
+          return _CODE_FORMAT_PROPERTY_CODE_39_ENABLED;
+        case CodeFormat.CODE_93:
+          return _CODE_FORMAT_PROPERTY_CODE_93_ENABLED;
+        case CodeFormat.CODE_128:
+          return _CODE_FORMAT_PROPERTY_CODE_128_ENABLED;
+        case CodeFormat.DATA_MATRIX:
+          return _CODE_FORMAT_PROPERTY_DATAMATRIX_ENABLED;
+        case CodeFormat.EAN_8:
+          return _CODE_FORMAT_PROPERTY_EAN_8_ENABLED;
+        case CodeFormat.EAN_13:
+          return _CODE_FORMAT_PROPERTY_EAN_13_ENABLED;
+        case CodeFormat.MAXICODE:
+          return _CODE_FORMAT_PROPERTY_MAXICODE_ENABLED;
+        case CodeFormat.PDF_417:
+          return _CODE_FORMAT_PROPERTY_PDF_417_ENABLED;
+        case CodeFormat.QR_CODE:
+          return _CODE_FORMAT_PROPERTY_QR_CODE_ENABLED;
+        case CodeFormat.RSS_14:
+          return _CODE_FORMAT_PROPERTY_RSS_ENABLED;
+        case CodeFormat.RSS_EXPANDED:
+          return _CODE_FORMAT_PROPERTY_RSS_EXPANDED_ENABLED;
+        case CodeFormat.UPC_A:
+          return _CODE_FORMAT_PROPERTY_UPC_A_ENABLE;
+        case CodeFormat.UPC_E:
+          return _CODE_FORMAT_PROPERTY_UPC_E_ENABLED;
         default:
       }
-    }
-    catch(e)
-    {
+    } catch (e) {
       print(e);
     }
     return null;
   }
 
-  Map<String, dynamic> getFormatsAsProperties(final List<CodeFormat> codeFormats) {
-    if(codeFormats == null || codeFormats.isEmpty) return {};
+  static CodeFormat valueOf(String name) => CodeFormat.values
+      .firstWhere((value) => value.name == name, orElse: () => null);
 
-    Map<String, dynamic> mapProperties = {};
-    codeFormats.forEach((codeFormat){
-      mapProperties[getPropertyName(codeFormat)] = true;
-    });
-    CodeFormat.values.forEach((codeFormat){
-      String propertyName = getPropertyName(codeFormat);
-      if(propertyName != null) mapProperties[propertyName] = mapProperties.containsKey(propertyName);
-    });
-    return mapProperties;
+  /// This function is Deprecated, [codeFormat.propertyName] instead
+  @deprecated
+  String getPropertyName(CodeFormat value) {
+    if (value == null) return null;
+    try {
+      switch (value) {
+        case CodeFormat.AZTEC:
+          return _CODE_FORMAT_PROPERTY_AZTEC_ENABLED;
+        case CodeFormat.CODABAR:
+          return _CODE_FORMAT_PROPERTY_CODABAR_ENABLED;
+        case CodeFormat.CODE_39:
+          return _CODE_FORMAT_PROPERTY_CODE_39_ENABLED;
+        case CodeFormat.CODE_93:
+          return _CODE_FORMAT_PROPERTY_CODE_93_ENABLED;
+        case CodeFormat.CODE_128:
+          return _CODE_FORMAT_PROPERTY_CODE_128_ENABLED;
+        case CodeFormat.DATA_MATRIX:
+          return _CODE_FORMAT_PROPERTY_DATAMATRIX_ENABLED;
+        case CodeFormat.EAN_8:
+          return _CODE_FORMAT_PROPERTY_EAN_8_ENABLED;
+        case CodeFormat.EAN_13:
+          return _CODE_FORMAT_PROPERTY_EAN_13_ENABLED;
+        case CodeFormat.MAXICODE:
+          return _CODE_FORMAT_PROPERTY_MAXICODE_ENABLED;
+        case CodeFormat.PDF_417:
+          return _CODE_FORMAT_PROPERTY_PDF_417_ENABLED;
+        case CodeFormat.QR_CODE:
+          return _CODE_FORMAT_PROPERTY_QR_CODE_ENABLED;
+        case CodeFormat.RSS_14:
+          return _CODE_FORMAT_PROPERTY_RSS_ENABLED;
+        case CodeFormat.RSS_EXPANDED:
+          return _CODE_FORMAT_PROPERTY_RSS_EXPANDED_ENABLED;
+        case CodeFormat.UPC_A:
+          return _CODE_FORMAT_PROPERTY_UPC_A_ENABLE;
+        case CodeFormat.UPC_E:
+          return _CODE_FORMAT_PROPERTY_UPC_E_ENABLED;
+        default:
+      }
+    } catch (e) {
+      print(e);
+    }
+    return null;
   }
 
+  /// This function is Deprecated, [codeFormat.name] instead
+  @deprecated
   String nameOf(CodeFormat value) {
     return value != null ? value.toString().split(".")[1] : null;
   }
 
+  /// Returns a Map of Honeywell's Barcode formats properties enabled by default
+  /// according to the List of CodeFormat specified.
+  /// [codeFormats] the List of CodeFormat enums to be converted to Honeywell properties
+  /// IMPORTANT:
+  /// - The codeFormats not specified in the [codeFormats] list will be set to disabled
+  /// - Empty codeFormats list means no properties at all
+  /// This function is Deprecated, use [getAsPropertiesComplement(...)] instead
+  @deprecated
+  static Map<String, dynamic> getFormatsAsProperties(
+      final List<CodeFormat> codeFormats) {
+    if (codeFormats?.isEmpty ?? true) return {};
+
+    Map<String, dynamic> mapProperties = {};
+    codeFormats.forEach((codeFormat) {
+      mapProperties[codeFormat?.propertyName] = true;
+    });
+    CodeFormat.values.forEach((codeFormat) {
+      String propertyName = codeFormat?.propertyName;
+      if (propertyName != null)
+        mapProperties[propertyName] = mapProperties.containsKey(propertyName);
+    });
+    return mapProperties;
+  }
+
+  /// Returns a Map of Honeywell's Barcode formats properties according to the
+  /// List of CodeFormat specified and the [enabled] value which is true by default.
+  /// [codeFormats] the List of CodeFormat enums to be converted to Honeywell properties
+  /// [enabled] the value to be set to the format property, true or false
+  static Map<String, dynamic> getAsProperties(final List<CodeFormat> codeFormats,
+      {bool enabled = true}) {
+    if (codeFormats?.isEmpty ?? true) return {};
+    enabled ??= true;
+    Map<String, dynamic> mapProperties = {};
+    codeFormats.forEach(
+        (codeFormat) => mapProperties[codeFormat?.propertyName] = enabled);
+    return mapProperties;
+  }
+
+  /// Returns a Map of Honeywell's Barcode formats properties according to the
+  /// List of CodeFormat specified and the [enabled] value which is true by default.
+  /// This function is useful when you want to enable some codeFormats and
+  /// disable the rest and the other way around.
+  /// [codeFormats] the List of CodeFormat enums to be converted to Honeywell properties
+  /// [enabled] the value to be set to the format property, true or false.
+  /// IMPORTANT:
+  /// - The codeFormats not specified in the [codeFormats] list will be set to the opposite of enabled like !enabled
+  static Map<String, dynamic> getAsPropertiesComplement(List<CodeFormat> codeFormats,
+      {bool enabled = true}) {
+    codeFormats ??= [];
+    enabled ??= true;
+
+    Map<String, dynamic> mapProperties =
+        getAsProperties(codeFormats, enabled: enabled);
+    CodeFormat.values.forEach((codeFormat) {
+      String propertyName = codeFormat?.propertyName;
+      if (propertyName != null && !mapProperties.containsKey(propertyName))
+        mapProperties[propertyName] = !enabled;
+    });
+    return mapProperties;
+  }
 }