+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-## 1.1.0+4 (2020-10-12)
+## 2.0.0+4 (2020-10-12)
 ### Changed
 - **stopScanner** function restores BarcodeScanner to initial properties, this means any custom properties will be replaced with the properties before **startScanner**
+- Honeywell sdk lib AAR updated to v1.00.00.0102
+- Example project updated to show a more detailed how to use.
+- CodeFormatUtils is now an extension of CodeFormat *(breaking change)*
 
 ## 1.0.1+3 (2020-10-10)
 ### Fixed